only remove tree if it exists ;-)
git-svn-id: http://svn.gna.org/svn/angel-app/trunk@942 82c71e93-f314-0410-b1e0-bdb1255beee9

+++ angel-app/src/angel_app/admin/directories.py
@@ -31,5 +31,7 @@ def removeDirectory(name):
     dirs = {
             'repository' : AngelConfig.get("common", "repository")
             }
-    shutil.rmtree(dirs[name], ignore_errors = False)
+    # only try to remove if it exists ;-)
+    if os.path.exists(dirs[name]):
+        shutil.rmtree(dirs[name], ignore_errors = False)
